+++ packages/perseus-editor/src/diffs/item-diff.tsx
@@ -4,31 +4,21 @@
  */
 
 import {Dependencies, type PerseusDependenciesV2} from "@khanacademy/perseus";
-import PropTypes from "prop-types";
 import * as React from "react";
 import _ from "underscore";
 
+import {AnswerAreaDiff} from "./answer-area-diff";
 import RendererDiff from "./renderer-diff";
-import WidgetDiff from "./widget-diff";
 
-const itemProps = PropTypes.shape({
-    question: PropTypes.shape({}).isRequired,
-    answerArea: PropTypes.shape({}).isRequired,
-    hints: PropTypes.arrayOf(PropTypes.any).isRequired,
-});
+import type {PerseusItem} from "@khanacademy/perseus-core";
 
 interface Props {
-    after: any;
-    before: any;
+    after: PerseusItem;
+    before: PerseusItem;
     dependencies: PerseusDependenciesV2;
 }
 
 class ItemDiff extends React.Component<Props> {
-    static propTypes = {
-        after: itemProps.isRequired,
-        before: itemProps.isRequired,
-    };
-
     render(): React.ReactNode {
         const {before, after} = this.props;
 
@@ -44,13 +34,18 @@ class ItemDiff extends React.Component<Props> {
             />
         );
 
-        const extras = (
-            <WidgetDiff
-                before={before.answerArea}
-                after={after.answerArea}
-                title="Question extras"
-            />
-        );
+        let extras: React.ReactNode;
+        // There might be a better way to do this, but for now I think it's fine to not
+        // show the diff if the answer area is set in both before and after.
+        if (before.answerArea && after.answerArea) {
+            extras = (
+                <AnswerAreaDiff
+                    before={before.answerArea}
+                    after={after.answerArea}
+                    title="Question extras"
+                />
+            );
+        }
 
         const hints = _.times(hintCount, function (n) {
             return (
@@ -74,8 +69,7 @@ class ItemDiff extends React.Component<Props> {
                 <div className="framework-perseus">
                     {question}
                     {extras}
-                    {/* eslint-disable-next-line @typescript-eslint/strict-boolean-expressions */}
-                    {hints && <div className="diff-separator" />}
+                    {hints.length > 0 && <div className="diff-separator" />}
                     {hints}
                 </div>
             </Dependencies.DependenciesContext.Provider>