+++ packages/perseus-score/src/index.ts
@@ -35,9 +35,11 @@ export {
     inputNumberAnswerTypes,
 } from "./widgets/input-number/score-input-number";
 
-export {scorePerseusItem, scoreWidgetsFunctional, flattenScores} from "./score";
-export {emptyWidgetsFunctional} from "./validate";
+export {scorePerseusItem, scoreWidgetsFunctional} from "./score";
+export {default as flattenScores} from "./util/flatten-scores";
+export {validateUserInput, emptyWidgetsFunctional} from "./validate";
 export {default as hasEmptyDINERWidgets} from "./has-empty-diner-widgets";
+export {default as isWidgetScoreable} from "./util/is-widget-scoreable";
 
 export type {
     PerseusMockWidgetRubric,