+++ packages/perseus/src/widgets/radio/__tests__/radio.testdata.ts
@@ -30,7 +30,6 @@ export const question: PerseusRenderer = generateTestPerseusRenderer({
                     generateRadioChoice(
                         "No value of $x$ satisfies the equation.",
                         {
-                            correct: false,
                             rationale: "$8$ satisfies the equation.",
                         },
                     ),
@@ -297,7 +296,6 @@ export const multiChoiceQuestionSimple: PerseusRenderer =
                             rationale: "Some people like to say Goodbye.",
                         }),
                         generateRadioChoice("None of these", {
-                            correct: false,
                             isNoneOfTheAbove: true,
                         }),
                     ],